cleanup(coral-tokens): refactor shades generation methods and file names
