+++ packages/coral-tokens/src/tokens/color/shades.json
@@ -8,62 +8,44 @@
       "10": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.1}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.1}"
+        }
       },
       "20": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.2}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.2}"
+        }
       },
       "30": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.3}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.3}"
+        }
       },
       "40": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.4}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.4}"
+        }
       },
       "50": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.5}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.5}"
+        }
       },
       "60": {
         "value": "{color.primary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.6}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.6}"
+        }
       }
     },
     "secondary": {
@@ -74,62 +56,44 @@
       "10": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.1}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.1}"
+        }
       },
       "20": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.2}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.2}"
+        }
       },
       "30": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.3}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.3}"
+        }
       },
       "40": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.4}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.4}"
+        }
       },
       "50": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.5}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.5}"
+        }
       },
       "60": {
         "value": "{color.secondary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.6}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.6}"
+        }
       }
     },
     "tertiary": {
@@ -140,62 +104,44 @@
       "10": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.1}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.1}"
+        }
       },
       "20": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.2}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.2}"
+        }
       },
       "30": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.3}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.3}"
+        }
       },
       "40": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.4}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.4}"
+        }
       },
       "50": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.5}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.5}"
+        }
       },
       "60": {
         "value": "{color.tertiary.0}",
         "category": "color",
-        "modify": [
-          {
-            "type": "set",
-            "amount": "{lightness.6}"
-          }
-        ]
+        "modify": {
+          "amount": "{lightness.6}"
+        }
       }
     }
   }
